优化

+++ src/views/account-manage/index.vue
@@ -236,7 +236,7 @@ const getList = (pageNo, pageSize) => {
     .then((res) => {
       const { code, data, page, message } = res;
       if (code === 200) {
-        const { list, page} = data;
+        const { list, page } = data;
         listData.value = list;
         console.log(data);
         setPage({ ...page });
@@ -262,21 +262,29 @@ const onAdd = () => {
 // 删除
 const onDel = (showLoading = true) => {
   const ids = selectList.value.map((e) => e.userId);
-
-  delLoading.value = true;
-  Api.DeleteAcc(ids)
-    .then((res) => {
-      const { code, message } = res;
-      if (code === 200) {
-        ElMessage.success("删除成功");
-        getList(1, page.pageSize);
-        clearSelect();
-      } else {
-        ElMessage.error(message);
-      }
+  const names = selectList.value.map((e) => e.realname);
+  
+  ElMessageBox.confirm(`确定删除用户【${names}】吗？`, "温馨提示", {
+    confirmButtonText: "确定",
+    type: "warning",
+  })
+    .then(() => {
+      delLoading.value = true;
+      Api.DeleteAcc(ids)
+        .then((res) => {
+          const { code, message } = res;
+          if (code === 200) {
+            ElMessage.success("删除成功");
+            getList(1, page.pageSize);
+            clearSelect();
+          } else {
+            ElMessage.error(message);
+          }
+        })
+        .catch((err) => console.log(err))
+        .finally(() => (delLoading.value = false));
     })
-    .catch((err) => console.log(err))
-    .finally(() => (delLoading.value = false));
+    .catch((err) => console.log(err));
 };
 
 // 编辑
@@ -290,18 +298,7 @@ const onEdit = (row, index) => {
 const onDelRow = (row, index) => {
   clearSelect();
   selectList.value.push(row);
-
-  ElMessageBox.confirm(`确定删除用户【${row.realname}】吗？`,
-      "温馨提示",
-      {
-        confirmButtonText: "确定",
-        type: "warning"
-      }
-    )
-      .then(() => {
-        onDel(false);
-      })
-      .catch(err => console.log(err));
+  onDel(false);
 };
 
 // 设置表格样式
